refactor(council): ADR-046 P0 — split council.py below the review cap (#408)
Verbatim moves, #380 playbook:
- council_stages.py (40K): stage1[_with_status], stage1.5 normalize,
  stage2, stage3, quick_synthesis, partial warning, title
- council_rankings.py (17K): parse_ranking_from_text, Borda aggregation,
  score-rank mismatch, shadow votes, label helpers
- council_usage.py (4K): ADR-012 timeout/status constants, ADR-011
  _add_cost_to_usage/_build_usage_summary
- council.py (44K, was 101K): orchestrators + config helpers +
  re-exports of every moved name

Back-compat: patches on llm_council.council.<stage_fn> still intercept
the orchestrators (they look up module globals in council.py); moved
code reads config helpers through the council module at CALL time, so
patch('llm_council.council.CHAIRMAN_MODEL', ...) semantics hold.
Patch sites whose consumers moved (query_* inside stages) retargeted to
llm_council.council_stages.* (5 test files).

Suite count identical (3041); every module <50K; ruff clean.

Closes #408
